+++ chpl/chpl-auth/src/main/java/gov/healthit/chpl/auth/entity/UserEntity.java
@@ -1,7 +1,5 @@
 package gov.healthit.chpl.auth.entity;
 
-import gov.healthit.chpl.auth.dto.UserPermissionDTO;
-
 import java.util.Date;
 import java.util.HashSet;
 import java.util.Set;
@@ -17,8 +15,7 @@ import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
-import org.hibernate.annotations.SQLDelete;
-import org.hibernate.annotations.Where;
+import gov.healthit.chpl.auth.dto.UserPermissionDTO;
 
 
 @Entity