+++ chpl/chpl-auth/src/main/java/gov/healthit/chpl/auth/manager/UserManager.java
@@ -1,6 +1,9 @@
 package gov.healthit.chpl.auth.manager;
 
 
+import java.util.List;
+import java.util.Set;
+
 import gov.healthit.chpl.auth.dto.UserDTO;
 import gov.healthit.chpl.auth.dto.UserPermissionDTO;
 import gov.healthit.chpl.auth.json.User;
@@ -10,9 +13,6 @@ import gov.healthit.chpl.auth.permission.UserPermissionRetrievalException;
 import gov.healthit.chpl.auth.user.UserCreationException;
 import gov.healthit.chpl.auth.user.UserManagementException;
 import gov.healthit.chpl.auth.user.UserRetrievalException;
-
-import java.util.List;
-import java.util.Set;
 
 public interface UserManager {