+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dao/CertifiedProductDAO.java
@@ -5,6 +5,7 @@ import java.util.List;
 
 import gov.healthit.chpl.dto.CertifiedProductDTO;
 import gov.healthit.chpl.dto.CertifiedProductDetailsDTO;
+import gov.healthit.chpl.entity.CertificationCriterionEntity;
 
 public interface CertifiedProductDAO {
 	
@@ -33,4 +34,5 @@ public interface CertifiedProductDAO {
 	
 	public List<CertifiedProductDTO> getByVersionIds(List<Long> versionIds);
 	public List<CertifiedProductDTO> getCertifiedProductsForDeveloper(Long vendorId);
+	public List<CertificationCriterionEntity> getRetiredTestTools(Long cpId);
 }