+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dao/impl/CertifiedProductDAOImpl.java
@@ -16,6 +16,7 @@ import gov.healthit.chpl.dao.EntityCreationException;
 import gov.healthit.chpl.dao.EntityRetrievalException;
 import gov.healthit.chpl.dto.CertifiedProductDTO;
 import gov.healthit.chpl.dto.CertifiedProductDetailsDTO;
+import gov.healthit.chpl.entity.CertificationCriterionEntity;
 import gov.healthit.chpl.entity.CertifiedProductDetailsEntity;
 import gov.healthit.chpl.entity.CertifiedProductEntity;
 
@@ -386,6 +387,21 @@ public class CertifiedProductDAOImpl extends BaseDAOImpl implements CertifiedPro
 			dtoResults.add(new CertifiedProductDetailsDTO(result));
 		}
 		return dtoResults;
+	}
+	
+	@Transactional(readOnly=true)
+	public List<CertificationCriterionEntity> getRetiredTestTools(Long cpId) {
+		Query query = entityManager.createQuery(
+				"SELECT c FROM CertificationCriterionEntity c "
+				+ "LEFT JOIN c.certificationResult cr "
+				+ "LEFT JOIN cr.certifiedProduct cp "
+				+ "LEFT JOIN cr.certificationResultTestTool crtt "
+				+ "LEFT JOIN crtt.testTool tt "
+				+ "WHERE (tt.retired = true AND CAST(cp.icsCode as integer) >= 1) AND c.deleted = false AND cp.id = :cpId)"
+				, CertificationCriterionEntity.class);
+		query.setParameter("cpId", cpId);
+		List<CertificationCriterionEntity> result = query.getResultList();
+		return result;
 	}
 	
 	@Transactional(readOnly=false)