+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/web/controller/CertifiedProductController.java
@@ -16,7 +16,6 @@ import org.apache.commons.csv.CSVParser;
 import org.apache.commons.csv.CSVRecord;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.aspectj.lang.annotation.Aspect;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -32,6 +31,7 @@ import org.springframework.web.multipart.MultipartFile;
 import com.fasterxml.jackson.core.JsonProcessingException;
 
 import gov.healthit.chpl.auth.Util;
+import gov.healthit.chpl.dao.CertifiedProductDAO;
 import gov.healthit.chpl.dao.EntityCreationException;
 import gov.healthit.chpl.dao.EntityRetrievalException;
 import gov.healthit.chpl.domain.ActivityConcept;
@@ -54,6 +54,7 @@ import gov.healthit.chpl.dto.CertifiedProductDetailsDTO;
 import gov.healthit.chpl.dto.CertifiedProductQmsStandardDTO;
 import gov.healthit.chpl.dto.CertifiedProductTargetedUserDTO;
 import gov.healthit.chpl.dto.PendingCertifiedProductDTO;
+import gov.healthit.chpl.entity.CertificationCriterionEntity;
 import gov.healthit.chpl.entity.CertificationStatusType;
 import gov.healthit.chpl.entity.PendingCertifiedProductEntity;
 import gov.healthit.chpl.manager.ActivityManager;
@@ -73,7 +74,6 @@ import io.swagger.annotations.ApiOperation;
 @Api(value="certified-products")
 @RestController
 @RequestMapping("/certified_products")
-@Aspect
 public class CertifiedProductController {
 	
 	private static final Logger logger = LogManager.getLogger(CertifiedProductController.class);
@@ -85,6 +85,7 @@ public class CertifiedProductController {
 	@Autowired CertificationBodyManager acbManager;
 	@Autowired ActivityManager activityManager;
 	@Autowired CertifiedProductValidatorFactory validatorFactory;
+	@Autowired CertifiedProductDAO cpDao;
 
 	@ApiOperation(value="List all certified products", 
 			notes="Default behavior is to return all certified products in the system. "
@@ -186,6 +187,24 @@ public class CertifiedProductController {
 					updateRequest.getErrorMessages().add("The CHPL Product Number has changed. The new CHPL Product Number " + updateRequest.getChplProductNumber() + " must be unique among all other certified products but one already exists with the same ID.");
 				}
 			} catch(EntityRetrievalException ex) {}
+		}
+		
+		List<CertificationCriterionEntity> retiredTestTools = cpDao.getRetiredTestTools(existingProduct.getId());
+		Boolean hasRetiredTestTool = false;
+		if(retiredTestTools.size() > 0){
+			hasRetiredTestTool = true;
+		}
+		
+		Boolean updateIcs = updateRequest.getIcs();
+		Boolean existingProductIcs = existingProduct.getIcs();
+		if(updateIcs == false && existingProductIcs == true && hasRetiredTestTool == true){
+			List<String> retiredTestToolNumbers = new ArrayList<String>();
+			for(CertificationCriterionEntity cce : retiredTestTools){
+				retiredTestToolNumbers.add(cce.getNumber());
+			}
+			
+			updateRequest.getErrorMessages().add("Cannot set ICS to false for a Certified Product with ICS=true and attested criteria that have a retired Test Tool. "
+					+ "The following are attested criteria that have a retired Test Tool: " + retiredTestToolNumbers.toString());
 		}
 		
 		if(updateRequest.getErrorMessages() != null && updateRequest.getErrorMessages().size() > 0) {