+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/manager/impl/CertifiedProductSearchManagerTest.java
@@ -15,7 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.github.springtestdbunit.DbUnitTestExecutionListener;
 import com.github.springtestdbunit.annotation.DatabaseSetup;
 
-import gov.healthit.chpl.caching.CacheInvalidationRule;
+import gov.healthit.chpl.caching.UnitTestRules;
 import gov.healthit.chpl.domain.CertifiedProductSearchResult;
 import gov.healthit.chpl.domain.SearchRequest;
 import gov.healthit.chpl.domain.SearchResponse;
@@ -36,7 +36,7 @@ public class CertifiedProductSearchManagerTest extends TestCase {
 	private CertifiedProductSearchManager certifiedProductSearchManager;
 	@Rule
     @Autowired
-    public CacheInvalidationRule cacheInvalidationRule;
+    public UnitTestRules cacheInvalidationRule;
 	
 	@Test
 	@Transactional